requirements.txt
pyserial
requests
airbrake==2.1.0